Implement bitop::bits_extract()
Extract bits from unsigned integer at the corresponding bit locations specified
by the mask to contiguous low bits of the result; the remaining upper bits in
the result are set to zero.

Matches the logic of PEXT - Parallel Bits Extract from BMI2 ISA, and on platforms
supporting BMI2 is implemented with this instruction.
When no BMI2 present, implemented serially but still quite efficient.